adding AI
