+++ package.json
@@ -18,7 +18,7 @@
     "clsx": "^2.0.0",
     "lucide-react": "^0.294.0",
     "next": "14.0.3",
-    "openai": "^4.20.1",
+    "openai": "^4.21.0",
     "react": "^18",
     "react-dom": "^18",
     "react-dropzone": "^14.2.3",
