--- src/components/CalendlyEmbed.js
@@ -1,25 +1,0 @@
-"use client"
-import React, { useEffect } from "react";
-
-const CalendlyEmbed = ({ url }) => {
-  useEffect(() => {
-    const head = document.querySelector("head");
-    const script = document.createElement("script");
-    script.setAttribute(
-      "src",
-      "https://assets.calendly.com/assets/external/widget.js"
-    );
-    head.appendChild(script);
-  }, []);
-
-  return (
-    <div
-      className="calendly-inline-widget h-full w-[100%]"
-      data-url={url}
-    >
-    
-    </div>
-  );
-};
-
-export default CalendlyEmbed;
